requirements.txt
selenium>=4.16.0
python-dotenv>=1.0.0
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.2
requests>=2.31.0 
flask>=2.3.2
